Backend HTTPS

+++ cmd/ingress/service_reconciler.go
@@ -18,9 +18,11 @@ const (
 	labelIngressInstance       = "ingress.lds.li/instance"
 	annMode                    = "ingress.lds.li/mode"
 	annSNIHostnames            = "ingress.lds.li/sni-hostnames"
+	annHTTPHostnames           = "ingress.lds.li/hostnames"
 	annProxyProtocol           = "ingress.lds.li/proxy-protocol"
 	modeTLSPassthrough         = "tls-passthrough"
 	modeTLSTermination         = "tls-termination"
+	modeHTTPS                  = "https"
 	proxyProtocolVersion1Value = "v1"
 )
 
@@ -50,12 +52,16 @@ func (s *ServiceReconciler) Reconcile(ctx context.Context, req reconcile.Request
 	}
 
 	mode := svc.Annotations[annMode]
-	if mode != modeTLSPassthrough && mode != modeTLSTermination {
+	if mode != modeTLSPassthrough && mode != modeTLSTermination && mode != modeHTTPS {
 		s.rdb.RemoveRoute(req.NamespacedName)
 		return reconcile.Result{}, nil
 	}
 
-	hostnames := splitCSV(svc.Annotations[annSNIHostnames])
+	hostnamesAnn := annSNIHostnames
+	if mode == modeHTTPS {
+		hostnamesAnn = annHTTPHostnames
+	}
+	hostnames := splitCSV(svc.Annotations[hostnamesAnn])
 	if len(hostnames) == 0 {
 		s.rdb.RemoveRoute(req.NamespacedName)
 		return reconcile.Result{}, nil
@@ -68,9 +74,7 @@ func (s *ServiceReconciler) Reconcile(ctx context.Context, req reconcile.Request
 
 	targetAddr := net.JoinHostPort(svc.Spec.ClusterIP, strconv.Itoa(int(svc.Spec.Ports[0].Port)))
 	proxyProto := svc.Annotations[annProxyProtocol] == proxyProtocolVersion1Value
-	terminateTLS := mode == modeTLSTermination
-
-	if err := s.rdb.SetRoute(req.NamespacedName, hostnames, targetAddr, proxyProto, terminateTLS); err != nil {
+	if err := s.rdb.SetRoute(req.NamespacedName, hostnames, targetAddr, mode, proxyProto); err != nil {
 		return reconcile.Result{}, err
 	}
 	return reconcile.Result{}, nil
